Bugfix and improve Security

+++ src/controllers/videoController.js
@@ -11,11 +11,18 @@ export const watch = async (req, res) => {
 	if (!video) {
 		return res.status("404").render("404", { pageTitle: "This Video dosen't exists."})
 	}
+	console.log(video);
 	res.render("watch", { pageTitle: video.title, video });
 }
 export const getEdit = async (req, res) => {
-	const { id } = req.params;
+	const { 
+		params: { id },
+		session: { user: { _id } } 
+	} = req;
 	const video = await Video.findById(id);
+	if(String(video.owner) !== String(_id)) {
+		return res.status("403").redirect("/");
+	}
 	return res.render("edit", { pageTitle: `Edit ${video.title}`, video });
 };
 export const postEdit = async (req, res) => {
@@ -76,7 +83,19 @@ export const postUpload = async (req, res) => {
 	*/
 
 export const deleteVideo = async (req, res) => {
-	const { id } = req.params;
+	const { 
+		params: { id },
+		session: { 
+			user: { _id }
+		}
+	} = req;
+	const video = Video.findById(id);
+	if (!video) {
+		return res.status("404").render("404", { pageTitle: "This Video dosen't exists."})
+	}
+	if(Stirng(video.owner) !== Stirng(_id)) {
+		return res.status("403").redirect("/");
+	}
 	await Video.findByIdAndDelete(id);
 	return res.redirect("/");
 };
